requirements.txt
flask
jinja2
requests
flask-session
python-dotenv
